오픈마켓 II [STEP 1] 웡빙, 보리사랑

+++ OpenMarket/OpenMarket/Extensions/Double+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  Int+Extensions.swift
+//  Double+Extensions.swift
 //  OpenMarket
 //
 //  Created by 웡빙, 보리사랑 on 2022/07/21.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func adoptDecimalStyle() -> String {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
